Replace `int` with `int16_t` or variations.

+++ src/Devices/EEPROM_TC.h
@@ -1,22 +1,23 @@
 #pragma once
 
 #include "Devices/EEPROM_TC.h"
+#include "TC_util.h"
 
 class EEPROM_TC {
 public:
   static EEPROM_TC* instance();
 
   // read and write
-  float eepromReadFloat(int address);
-  int eepromReadInt(int address);
-  void eepromWriteFloat(int address, float value);
-  void eepromWriteInt(int address, int value);
+  float eepromReadFloat(uint16_t address);
+  int32_t eepromReadInt(uint16_t address);
+  void eepromWriteFloat(uint16_t address, float value);
+  void eepromWriteInt(uint16_t address, int32_t value);
 
   // accessor methods
   float getAmplitude();
   float getCorrectedTemp();
   float getFrequency();
-  int getGoogleSheetInterval();
+  uint16_t getGoogleSheetInterval();
   float getGranularity();
   bool getHeat();
   float getKD();
@@ -29,7 +30,7 @@ public:
   float getPHInterval();
   float getPHSeriesPointer();
   float getPHSeriesSize();
-  int getTankID();
+  uint16_t getTankID();
   float getTemp();
   float getTempDelay();
   float getTempInterval();
@@ -40,7 +41,7 @@ public:
   void setAmplitude(float value);
   void setCorrectedTemp(float value);
   void setFrequency(float value);
-  void setGoogleSheetInterval(int value);
+  void setGoogleSheetInterval(uint16_t value);
   void setGranularity(float value);
   void setHeat(bool value);
   void setKD(float value);
@@ -53,7 +54,7 @@ public:
   void setPHInterval(float value);
   void setPHSeriesPointer(float value);
   void setPHSeriesSize(float value);
-  void setTankID(int value);
+  void setTankID(uint16_t value);
   void setTemp(float value);
   void setTempDelay(float value);
   void setTempInterval(float value);
@@ -64,30 +65,30 @@ public:
 
 private:
   // instance variables from v0.197
-  const int PH_ADDRESS = 0;          // 9.999
-  const int TEMP_ADDRESS = 4;        // 99.99
-  const int TANK_ID_ADDRESS = 8;     // 999
-  const int TEMP_CORR_ADDRESS = 12;  // 99.99
-  const int KP_ADDRESS = 20;         // float
-  const int KI_ADDRESS = 28;         // float
-  const int KD_ADDRESS = 36;         // float
-  const int MAC_ADDRESS = 44;        // 8 byte
-  const int HEAT_ADDRESS = 52;       // bool
+  const uint16_t PH_ADDRESS = 0;          // 9.999
+  const uint16_t TEMP_ADDRESS = 4;        // 99.99
+  const uint16_t TANK_ID_ADDRESS = 8;     // 999
+  const uint16_t TEMP_CORR_ADDRESS = 12;  // 99.99
+  const uint16_t KP_ADDRESS = 20;         // float
+  const uint16_t KI_ADDRESS = 28;         // float
+  const uint16_t KD_ADDRESS = 36;         // float
+  const uint16_t MAC_ADDRESS = 44;        // 8 byte
+  const uint16_t HEAT_ADDRESS = 52;       // bool
   // new with v0.2
-  const int AMPLITUDE_ADDRESS = 56;
-  const int FREQUENCY_ADDRESS = 60;
-  const int GRANULARITY_ADDRESS = 64;   // granularity for SD logging interval
-  const int MAX_DATA_AGE_ADDRESS = 68;  // max data age for SD card
-  const int PH_SERIES_SIZE_ADDRESS = 72;
-  const int PH_SERIES_POINTER_ADDRESS = 76;
-  const int TEMP_SERIES_SIZE_ADDRESS = 80;
-  const int TEMP_SERIES_POINTER_ADDRESS = 84;
-  const int PH_INTERVAL_ADDRESS = 88;
-  const int PH_DELAY_ADDRESS = 92;
-  const int TEMP_INTERVAL_ADDRESS = 96;
-  const int TEMP_DELAY_ADDRESS = 100;
+  const uint16_t AMPLITUDE_ADDRESS = 56;
+  const uint16_t FREQUENCY_ADDRESS = 60;
+  const uint16_t GRANULARITY_ADDRESS = 64;   // granularity for SD logging interval
+  const uint16_t MAX_DATA_AGE_ADDRESS = 68;  // max data age for SD card
+  const uint16_t PH_SERIES_SIZE_ADDRESS = 72;
+  const uint16_t PH_SERIES_POINTER_ADDRESS = 76;
+  const uint16_t TEMP_SERIES_SIZE_ADDRESS = 80;
+  const uint16_t TEMP_SERIES_POINTER_ADDRESS = 84;
+  const uint16_t PH_INTERVAL_ADDRESS = 88;
+  const uint16_t PH_DELAY_ADDRESS = 92;
+  const uint16_t TEMP_INTERVAL_ADDRESS = 96;
+  const uint16_t TEMP_DELAY_ADDRESS = 100;
   // new with v0.3
-  const int GOOGLE_INTERVAL_ADDRESS = 108;
+  const uint16_t GOOGLE_INTERVAL_ADDRESS = 108;
 
   // class variables
   static EEPROM_TC* _instance;
