+++ test/LiquidCrystalTest.cpp
@@ -18,16 +18,16 @@ unittest(loop) {
   assertEqual(2, lines.size());
   assertEqual(16, lines.at(0).length());
   assertEqual(16, lines.at(1).length());
-  assertEqual(0, (int)lines.at(0).at(0));
-  assertEqual(1, (int)lines.at(0).at(1));
-  assertEqual(2, (int)lines.at(0).at(2));
-  assertEqual(3, (int)lines.at(0).at(3));
-  assertEqual(4, (int)lines.at(1).at(0));
-  assertEqual(5, (int)lines.at(1).at(1));
-  assertEqual(6, (int)lines.at(1).at(2));
-  assertEqual(7, (int)lines.at(1).at(3));
-  for (int i = 0; i < lines.size(); ++i) {
-    for (int j = 0; j < lines.at(i).size(); ++j) {
+  assertEqual(0, (uint16_t)lines.at(0).at(0));
+  assertEqual(1, (uint16_t)lines.at(0).at(1));
+  assertEqual(2, (uint16_t)lines.at(0).at(2));
+  assertEqual(3, (uint16_t)lines.at(0).at(3));
+  assertEqual(4, (uint16_t)lines.at(1).at(0));
+  assertEqual(5, (uint16_t)lines.at(1).at(1));
+  assertEqual(6, (uint16_t)lines.at(1).at(2));
+  assertEqual(7, (uint16_t)lines.at(1).at(3));
+  for (size_t i = 0; i < lines.size(); ++i) {
+    for (size_t j = 0; j < lines.at(i).size(); ++j) {
       if (lines.at(i).at(j) < ' ') {
         lines.at(i).at(j) = '?';
       }
